refactor(pages): split handlers, move router to mod.rs
