+++ server/src/openapi.rs
@@ -168,13 +168,13 @@ use utoipa::{Modify, OpenApi};
         crate::discovery::handlers::browse_guilds,
         crate::discovery::handlers::join_discoverable,
         // Guild Pages
-        crate::pages::handlers::list_guild_pages,
-        crate::pages::handlers::create_guild_page,
-        crate::pages::handlers::update_guild_page,
-        crate::pages::handlers::delete_guild_page,
-        crate::pages::handlers::reorder_guild_pages,
-        crate::pages::handlers::get_guild_page,
-        crate::pages::handlers::accept_guild_page,
+        crate::pages::guild_pages::list_guild_pages,
+        crate::pages::guild_pages::create_guild_page,
+        crate::pages::guild_pages::update_guild_page,
+        crate::pages::guild_pages::delete_guild_page,
+        crate::pages::guild_pages::reorder_guild_pages,
+        crate::pages::guild_pages::get_guild_page,
+        crate::pages::guild_pages::accept_guild_page,
         // Admin
         crate::admin::system::get_admin_status,
         crate::admin::system::get_admin_stats,
@@ -304,14 +304,14 @@ use utoipa::{Modify, OpenApi};
         crate::connectivity::get_sessions,
         crate::connectivity::get_session_detail,
         // Pages
-        crate::pages::handlers::list_platform_pages,
-        crate::pages::handlers::get_platform_page,
-        crate::pages::handlers::create_platform_page,
-        crate::pages::handlers::get_pending_acceptance,
-        crate::pages::handlers::reorder_platform_pages,
-        crate::pages::handlers::update_platform_page,
-        crate::pages::handlers::delete_platform_page,
-        crate::pages::handlers::accept_page,
+        crate::pages::platform::list_platform_pages,
+        crate::pages::platform::get_platform_page,
+        crate::pages::platform::create_platform_page,
+        crate::pages::platform::get_pending_acceptance,
+        crate::pages::platform::reorder_platform_pages,
+        crate::pages::platform::update_platform_page,
+        crate::pages::platform::delete_platform_page,
+        crate::pages::platform::accept_page,
         // Settings
         crate::api::settings::get_server_settings,
         crate::api::settings::get_upload_limits,
